+++ Assets/ConsolePro/Remote/LiteNetLib/LiteNetLib.csproj.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 36ae2e48e7a7343fcb1503f56b5d4485
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
