+++ Assets/Plugins/Borodar/RainbowHierarchy/Source/RainbowHierarchy.src.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 617b3f1032a08c14ebfedfa340767cdf
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
