+++ Assets/Scripts/Events/MainEventsManager.cs
@@ -1,0 +1,13 @@
+using System;
+using UnityEngine;
+
+namespace Main.Events
+{
+    public static class MainEventsManager
+    {
+        public static Action<Transform> rightHandTransformUpdate;
+        public static Action<Transform> leftHandTransformUpdate;
+        public static Action<Transform> headTransformUpdate;
+        public static Action<Transform> bodyTransformUpdate;
+    }
+}
