+++ Assets/Scripts/Events/MainEventsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b225e261405bc5428c23b9963d90199
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
