+++ Assets/Scripts/IK.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37369c7f7a7260440857e6ed7faa4a2f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
