+++ Assets/Scripts/Mimics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4971b35688a1cbc4e86d5a3905cd51f0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
