+++ Assets/Scripts/Mimics/MimicChar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f5d1a1ead7e69dd4eaabd88dccddc4e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
