+++ Assets/Scripts/Mimics/MimicReference.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 48610fb5646ab5f4c9cf2711ef68fb6d
+guid: 17ff56da1970cc74184107cff2867419
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
